interface de test et enregistrement de n'importe qu'elle extension de media

+++ scriber/src/ressources/Controller.java
@@ -68,7 +68,7 @@ public class Controller implements Initializable {
         exerciceLoader.loadMediaData(exerciceLoader.chargerMediaDepuisExercice(ExerciceFile.getPath()));
         titreMedia.setText(exerciceLoader.getTitle());
         albumMedia.setText(exerciceLoader.getAlbum());
-        annerMedia.setText( ((Integer) exerciceLoader.getYear()).toString());
+        annerMedia.setText( ((Float) exerciceLoader.getYear()).toString());
     }
 
 }
